requirements.txt
torch
torchvision
opencv-python
numpy
scikit-learn
matplotlib
pillow
tqdm